fix(domain): remove unused go-openai import in generate.go refactor(domain): relocate CreateClient function to invoke.go
