+++ tools/seltabls/domain/invoke.go
@@ -9,6 +9,15 @@ import (
 	"golang.org/x/sync/errgroup"
 )
 
+// CreateClient creates a new client for the given api key.
+func CreateClient(baseURL string, apiKey string) *openai.Client {
+	cfg := openai.DefaultConfig(apiKey)
+	cfg.BaseURL = baseURL
+	cfg.APIVersion = string(openai.APITypeOpenAI)
+	cfg.APIType = openai.APITypeOpenAI
+	return openai.NewClientWithConfig(cfg)
+}
+
 // InvokeJSONSimple is a function for generating json using the OpenAI API with the given prompt.
 //
 // It does not decode the json.
